remove dependency for tests
