+++ .travis.yml
@@ -25,8 +25,6 @@ install:
 
 script:
   - |
-    # update pip
-    pip install -U pip
     # install tox
     pip install tox
     # Run the tests
