+++ tox.ini
@@ -9,7 +9,6 @@ deps =
     codecov~=2.0
     hypothesis~=4.56
     pyserial~=3.0
-    cantact~=0.0.5
 
 commands =
     pytest
